# internal/http/handlers/category_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/qullDev/book_API/internal/domain/book"
	"github.com/qullDev/book_API/internal/domain/category"
	"gorm.io/gorm"
)

type CategoryHandler struct {
	db *gorm.DB
}

func NewCategoryHandler(db *gorm.DB) *CategoryHandler {
	return &CategoryHandler{db: db}
}

func (h *CategoryHandler) Register(rg *gin.RouterGroup) {
	rg.GET("", h.List)
	rg.POST("", h.Create)
	rg.GET("/:id", h.Detail)
	rg.PUT("/:id", h.Update)
	rg.DELETE("/:id", h.Delete)
	rg.GET("/:id/books", h.ListBooks)
}

type createCategoryReq struct {
	Name string `json:"name" binding:"required,min=1,max=100"`
}

type updateCategoryReq struct {
	Name string `json:"name" binding:"required,min=1,max=100"`
}

// @Summary List all categories
// @Description Get list of all categories
// @Tags categories
// @Security BearerAuth
// @Accept json
// @Produce json
// @Success 200 {object} map[string][]category.Category
// @Router /api/categories [get]
func (h *CategoryHandler) List(c *gin.Context) {
	var items []category.Category
	if err := h.db.Order("name asc").Find(&items).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"message": "gagal mengambil data kategori"})
		return
	}
	c.JSON(http.StatusOK, gin.H{"data": items})
}

// @Summary Create category
// @Description Create a new category
// @Tags categories
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param category body createCategoryReq true "Category data" example({"name": "Fiction"})
// @Success 201 {object} map[string]category.Category "example={'data':{'id':'550e8400-e29b-41d4-a716-446655440000','name':'Fiction','created_at':'2024-01-20T10:00:00Z'}}"
// @Failure 400 {object} gin.H "example={'message':'payload tidak valid'}"
// @Router /api/categories [post]
func (h *CategoryHandler) Create(c *gin.Context) {
	var req createCategoryReq
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"message": "payload tidak valid", "error": err.Error()})
		return
	}
	item := category.Category{Name: req.Name}
	if err := h.db.Create(&item).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"message": "gagal menambahkan kategori"})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"data": item})
}

// @Summary Get category detail
// @Description Get detail of a category
// @Tags categories
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path string true "Category ID" format(uuid)
// @Success 200 {object} map[string]category.Category
// @Failure 400,404 {object} gin.H
// @Router /api/categories/{id} [get]
func (h *CategoryHandler) Detail(c *gin.Context) {
	idStr := c.Param("id")
	id, err := uuid.Parse(idStr)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"message": "id tidak valid"})
		return
	}
	var item category.Category
	if err := h.db.First(&item, "id = ?", id).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			c.JSON(http.StatusNotFound, gin.H{"message": "kategori tidak ditemukan"})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"message": "gagal mengambil detail kategori"})
		return
	}
	c.JSON(http.StatusOK, gin.H{"data": item})
}

// @Summary Update category
// @Description Update an existing category
// @Tags categories
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path string true "Category ID" format(uuid)
// @Param category body updateCategoryReq true "Updated category data"
// @Success 200 {object} map[string]category.Category
// @Failure 400,404 {object} gin.H
// @Router /api/categories/{id} [put]
func (h *CategoryHandler) Update(c *gin.Context) {
	idStr := c.Param("id")
	id, err := uuid.Parse(idStr)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"message": "id tidak valid"})
		return
	}
	var req updateCategoryReq
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"message": "payload tidak valid", "error": err.Error()})
		return
	}
	var item category.Category
	if err := h.db.First(&item, "id = ?", id).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			c.JSON(http.StatusNotFound, gin.H{"message": "kategori tidak ditemukan"})
			return
		}
		c.JSON(http.StatusInternalServerError, gin.H{"message": "gagal mengambil data kategori"})
		return
	}
	item.Name = req.Name
	if err := h.db.Save(&item).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"message": "gagal mengupdate kategori"})
		return
	}
	c.JSON(http.StatusOK, gin.H{"data": item})
}

// @Summary Delete category
// @Description Delete a category
// @Tags categories
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path string true "Category ID" format(uuid)
// @Success 200 {object} gin.H
// @Failure 400,404 {object} gin.H
// @Router /api/categories/{id} [delete]
func (h *CategoryHandler) Delete(c *gin.Context) {
	idStr := c.Param("id")
	id, err := uuid.Parse(idStr)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"message": "id tidak valid"})
		return
	}
	res := h.db.Delete(&category.Category{}, "id = ?", id)
	if res.Error != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"message": "gagal menghapus kategori"})
		return
	}
	if res.RowsAffected == 0 {
		c.JSON(http.StatusNotFound, gin.H{"message": "kategori tidak tersedia"})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "kategori berhasil dihapus"})
}

// @Summary List books in category
// @Description Get list of books in a specific category
// @Tags categories
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path string true "Category ID" format(uuid)
// @Success 200 {object} map[string][]book.Book
// @Failure 400 {object} gin.H
// @Router /api/categories/{id}/books [get]
func (h *CategoryHandler) ListBooks(c *gin.Context) {
	idStr := c.Param("id")
	id, err := uuid.Parse(idStr)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"message": "id tidak valid"})
		return
	}
	var books []book.Book
	if err := h.db.Where("category_id = ?", id).Find(&books).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"message": "gagal mengambil buku pada kategori"})
		return
	}
	c.JSON(http.StatusOK, gin.H{"data": books})
}